+++ geoprox/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Improved default config behavior
+
 ## 0.4.2
 
 - Added support for environment variables ([#6](https://github.com/ezrasingh/geoprox/issues/6)).